// backend/src/types.ts
export type AnnotationID = UUID;

export type AnnotationLocation = {
  start: number;
  end: number;
};

export type Annotation = {
  id: AnnotationID;
  location: AnnotationLocation;
  content: string;
  author: UserID;
  createdTimestamp: ISO8601PreciseTimestamp;
  modifiedTimestamp: ISO8601PreciseTimestamp;
  isDeleted: boolean;
};

export type AnyRecipeStepInput = BaseRecipeStepInput &
  (RecipeStepConstantInput | RecipeStepArtifactInput | RecipeStepMetricInput);
// ...
export type ArtifactMetricGeneration = {
  eventSummaryID: EventSummaryID;
  recipeRunID: RecipeRunID;
  generationID: GenerationID;
  modelID: string;
  modelParams?: EvaluationModelParameters | undefined;
  endTimestamp: ISO8601PreciseTimestamp;
  didComplete: boolean;
};

export type ArtifactPathComponent = {
  kind?: (ArtifactKind | null) | undefined;
  id: ArtifactID;
};

export type ArtifactPathPatternComponent =
  | {
      kind: ArtifactKind;
      id?: (ArtifactID | null) | undefined;
    }
  | {
      id: ArtifactID;
    };

export type ArtifactPathPattern = ArtifactPathPatternComponent[];

export type ArtifactPathRecipeTrigger = {
  id: RecipeTriggerID;
  evaluationGroupID: EvaluationGroupID;
  name: string;
  kind: typeof RecipeTriggerKind.artifactPath;
  creationTimestamp: ISO8601PreciseTimestamp;
  updateTimestamp: ISO8601PreciseTimestamp;
  artifactPathPattern: ArtifactPathPattern;
};

export type ArtifactPath = ArtifactPathComponent[];

export type ArtifactPatternConfiguration = {
  pattern: ArtifactPathPattern;
  displayName: DisplayName;
  allowCreation?: boolean | undefined;
};

export type Artifact = {
  artifactPath: ArtifactPath;
  sourceArtifactPaths?: ArtifactPath[] | undefined;
  snapshots: ArtifactSnapshot[];
  metrics?: Metric[] | undefined;
  generations?: ArtifactMetricGeneration[] | undefined;
};

export type ArtifactSelector = {
  tags?: TagValue[] | undefined;
  artifactPath: ArtifactPath;
  eventSummaryIDs?: EventSummaryID[] | undefined;
  recipeRunIDs?: RecipeRunID[] | undefined;
  generationIDs?: GenerationID[] | undefined;
};
// ...
export type ArtifactSnapshotStrict = {
  artifactPath: ArtifactPath;
  sourceArtifactSelectors: ArtifactSelector[];
  eventSummaryID: EventSummaryID;
  tags: {
    [x: string]: string;
  };
  metadata: {
    [x: string]: string;
  };
  timestamp: ISO8601PreciseTimestamp;
  content: RenderedContent;
  metrics: MetricStrict[];
  generations: ArtifactMetricGeneration[];
  /** Mapping from annotation ID to annotation */
  annotations: {
    [x: string]: Annotation;
  };
  /** Mapping from MetricReviewID to MetricReview */
  reviews: {
    [x: string]: MetricReview;
  };
  /** Mapping from an evaluation (identified by EvalutionGroupID) to a due date */
  dueDates: {
    [x: string]: ISO8601PreciseTimestamp;
  };
};

export type ArtifactStrict = {
  artifactPath: ArtifactPath;
  snapshots: ArtifactSnapshotStrict[];
  metrics: MetricStrict[];
  generations: ArtifactMetricGeneration[];
};

export type BaseRecipeStepInput = {
  kind: RecipeStepInputKind;
  token: string;
};

export type BaseRecipeStepOutput = {
  kind: RecipeStepOutputKind;
  key: string;
};

export type BaseRecipeStep = {
  id: RecipeStepID;
  name: string;
  kind: RecipeStepKind;
  status: RecipeStepStatus;
  creationTimestamp: ISO8601PreciseTimestamp;
  updateTimestamp: ISO8601PreciseTimestamp;
  dependencies: RecipeStepID[];
};

export type BaseRecipeTrigger = {
  id: RecipeTriggerID;
  evaluationGroupID: EvaluationGroupID;
  name: string;
  kind: RecipeTriggerKind;
  creationTimestamp: ISO8601PreciseTimestamp;
  updateTimestamp: ISO8601PreciseTimestamp;
};
// ...
export type CopyRecipeStep = {
  id: RecipeStepID;
  name: string;
  kind: typeof RecipeStepKind.copy;
  status: RecipeStepStatus;
  creationTimestamp: ISO8601PreciseTimestamp;
  updateTimestamp: ISO8601PreciseTimestamp;
  dependencies: RecipeStepID[];
  inputs: RecipeStepMetricInput[];
  outputs: RecipeStepArtifactOutput[];
};
// ...
export type EvaluateRecipeStep = {
  id: RecipeStepID;
  name: string;
  kind: typeof RecipeStepKind.evaluate;
  status: RecipeStepStatus;
  creationTimestamp: ISO8601PreciseTimestamp;
  updateTimestamp: ISO8601PreciseTimestamp;
  dependencies: RecipeStepID[];
  inputs: AnyRecipeStepInput[];
  outputs: (BaseRecipeStepOutput & (RecipeStepSnapshotOutput | RecipeStepMetricOutput))[];
  promptTemplate: string;
  userPrompt: string;
  model: RecipeEvaluationModel;
};

export type EvaluationGroupID = UUID;

export type EvaluationModelConfiguration = {
  id: string;
  displayName: string;
  description: string;
  provider: string;
  costMultiplier: string;
  defaultParameters?: EvaluationModelParameters | undefined;
};

export type EvaluationModelParameters = {
  temperature?: number | undefined;
  topP?: number | undefined;
  maxNewTokens?: number | undefined;
};

export type EventSummaryID = UUID;

export type ExportMetricsRequest = {
  orgID: OrganizationID;
  metricSetID?: string | undefined;
};

export type Filter = PrimitiveValueFilter | KeyPathFilter | GroupFilter;

export type GenerationID = string;

export const GroupFilterOperator = {
  all: "all",
  any: "any",
  none: "none",
} as const;
export type GroupFilterOperator = (typeof GroupFilterOperator)[keyof typeof GroupFilterOperator];

export type GroupFilter = {
  operator: GroupFilterOperator;
  filters: Filter[];
};

export type ISO8601PreciseTimestamp = string | string;

export type KeyPathFilter = {
  operator: ValueFilterOperator;
  keyPath: KeyPath;
  value: PrimitiveValue;
};

export type KeyPath = string;

export type LoadRecipesRequest = {
  orgID: OrganizationID;
};

export type LoadRecipesResponse = {
  recipes: Recipe[];
  cancelledEvaluationGroupIDs: EvaluationGroupID[];
};

export type MetricDefinition = {
  id: MetricID;
  precursorID?: MetricID | undefined;
  isDeleted?: true | undefined;
  name: string;
  description?: string | undefined;
  order?: string | undefined;
  group?: string | undefined;
  relatedMetricsIDs?: MetricID[] | undefined;
  metricValueAggregationRule?: MetricValueAggregationRule | undefined;
  displayValues?: [PrimitiveValue, string][] | undefined;
  acceptanceValue?: Filter | undefined;
  rejectionValue?: Filter | undefined;
  kind?: ("string" | "number" | "boolean" | "icon") | undefined;
  unit?: string | undefined;
  color?: string | undefined;
};
// ...
export type MetricID = string;

export type MetricRecording = {
  eventSummaryID: EventSummaryID;
  recipeRunID?: RecipeRunID | undefined;
  generationID?: GenerationID | undefined;
  evaluationGroupID: EvaluationGroupID;
  value: PrimitiveValue;
  examples?: MetricExample[] | undefined;
};

export type MetricRecordingStrict = {
  eventSummaryID: EventSummaryID;
  recipeRunID: RecipeRunID;
  generationID: GenerationID;
  evaluationGroupID: EvaluationGroupID;
  value: PrimitiveValue;
  examples?: MetricExample[] | undefined;
};

export type MetricReviewID = UUID;

export type MetricReview = {
  id: MetricReviewID;
  metricId: MetricID;
  recipeRunId: RecipeRunID;
  evaluationGroupId: EvaluationGroupID;
  value: MetricReviewValue;
  author: UserID;
  createdTimestamp: ISO8601PreciseTimestamp;
  modifiedTimestamp: ISO8601PreciseTimestamp;
};

export const MetricReviewValue = {
  approved: "approved",
  denied: "denied",
  not_applicable: "not_applicable",
} as const;
export type MetricReviewValue = (typeof MetricReviewValue)[keyof typeof MetricReviewValue];

export type Metric = {
  id: MetricID;
  values: MetricRecording[];
  isMock?: boolean | undefined;
};

export type MetricStrict = {
  id: MetricID;
  values: MetricRecordingStrict[];
  isMock?: true | undefined;
};

export const MetricValueAggregationRule = {
  uniqueValues: "uniqueValues",
  uniformValues: "uniformValues",
  anyTrue: "anyTrue",
  allTrue: "allTrue",
  concatenate: "concatenate",
} as const;
export type MetricValueAggregationRule = (typeof MetricValueAggregationRule)[keyof typeof MetricValueAggregationRule];
// ...
export type OrganizationID = UUID;

export type Organization = {
  id: OrganizationID;
  name: string;
  template?: OrganizationTemplate | undefined;
  isDeleted?: true | undefined;
};

export const OrganizationTemplate = {
  demo: "demo",
  general: "general",
} as const;
export type OrganizationTemplate = (typeof OrganizationTemplate)[keyof typeof OrganizationTemplate];
// ...
export type PrimitiveValueFilter = {
  operator: ValueFilterOperator;
  value: PrimitiveValue;
};

export type PrimitiveValue = boolean | string | number | null;

export type RecipeEvaluationModel = {
  name: string;
  parameters: {
    temperature?: number | undefined;
    topP?: number | undefined;
    maxNewTokens?: number | undefined;
  };
};

export type RecipeID = string;

export type RecipeRunID = string;

export type Recipe = {
  id: RecipeID;
  name: string;
  description?: string | undefined;
  isDeleted?: true | undefined;
  creationTimestamp: ISO8601PreciseTimestamp;
  updateTimestamp: ISO8601PreciseTimestamp;
  triggers: RecipeTrigger[];
  steps: RecipeStep[];
};

export type RecipeStepArtifactInput = {
  kind: typeof RecipeStepInputKind.artifact;
  token: string;
  input: {
    childArtifactPath: ArtifactPath;
    keyPath: KeyPath;
  };
};

export type RecipeStepArtifactOutput = {
  kind: typeof RecipeStepOutputKind.artifact;
  key: string;
  output: {
    childArtifactPath: ArtifactPath;
    metricValueAggregationRule: MetricValueAggregationRule;
  };
};

export type RecipeStepConstantInput = {
  kind: typeof RecipeStepInputKind.constant;
  token: string;
  input: string;
};

export type RecipeStepID = string;

export const RecipeStepInputKind = {
  constant: "constant",
  artifact: "artifact",
  metric: "metric",
} as const;
export type RecipeStepInputKind = (typeof RecipeStepInputKind)[keyof typeof RecipeStepInputKind];

export const RecipeStepKind = {
  evaluate: "evaluate",
  copy: "copy",
} as const;
export type RecipeStepKind = (typeof RecipeStepKind)[keyof typeof RecipeStepKind];

export type RecipeStepMetricInput = {
  kind: typeof RecipeStepInputKind.metric;
  token: string;
  input: {
    childArtifactPath: ArtifactPath;
    metricID: MetricID;
  };
};

export type RecipeStepMetricOutput = {
  kind: typeof RecipeStepOutputKind.metric;
  key: string;
  output: {
    childArtifactPath: ArtifactPath;
    metricID: MetricID;
    includeEvidence: boolean;
  };
};

export const RecipeStepOutputKind = {
  artifact: "artifact",
  snapshot: "snapshot",
  metric: "metric",
} as const;
export type RecipeStepOutputKind = (typeof RecipeStepOutputKind)[keyof typeof RecipeStepOutputKind];

export type RecipeStep = BaseRecipeStep & (EvaluateRecipeStep | CopyRecipeStep);

export type RecipeStepSnapshotOutput = {
  kind: typeof RecipeStepOutputKind.snapshot;
  key: string;
  output: {
    childArtifactPath: ArtifactPath;
  };
};

export const RecipeStepStatus = {
  enabled: "enabled",
  disabled: "disabled",
  migrated: "migrated",
  hidden: "hidden",
} as const;
export type RecipeStepStatus = (typeof RecipeStepStatus)[keyof typeof RecipeStepStatus];

export type RecipeTriggerID = string;

export const RecipeTriggerKind = {
  artifactPath: "artifactPath",
} as const;
export type RecipeTriggerKind = (typeof RecipeTriggerKind)[keyof typeof RecipeTriggerKind];

export type RecipeTrigger = BaseRecipeTrigger & ArtifactPathRecipeTrigger;
// ...
export type RenderedContent =
  | PrimitiveValue
  | {
      [x: string]: RenderedContent;
    };
// ...
export type TagValue = {
  tag: string;
  value: string;
};

export type URL = string;

export type UUID = string;

export type UpdateArtifactSnapshotRequest = {
  orgID: OrganizationID;
  artifactPath: ArtifactPath;
  eventSummaryID: EventSummaryID;
  snapshotDelta: ArtifactSnapshotDelta;
};

export type UpdateArtifactSnapshotResponse = {
  status: "success" | "error";
  message: string;
};

export type UserID = string;